pyproject.toml
[project]
name = "git-digest"
version = "0.1.0"
description = "Fetch latest git updates (commits, releases) from configured repos and generate digests for cron."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "GitPython>=3.1.40",
    "PyYAML>=6.0",
    "python-dotenv>=1.0",
    "requests>=2.28",
    "fastmcp>=2,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[project.scripts]
git-digest = "git_updates.cli:main"
git-digest-mcp = "git_updates.mcp:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/git_updates"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
